Debug for TPU

+++ src/fno_vmc/fno_jax/spectral_conv.py
@@ -22,7 +22,7 @@ class SpectralConv2d(nn.Module):
     def __call__(self, x):  # x: (batch, Nx, Ny, in_channels)
         batch, Nx, Ny, _ = x.shape
         # 转到频域
-        x_ft = rfft2d(x.transpose((0,3,1,2)))  # -> (batch, in_channels, Nx, Ny//2+1)
+        x_ft = rfft2d(x.astype(jnp.float32).transpose((0,3,1,2))).astype(jnp.complex64) # -> (batch, in_channels, Nx, Ny//2+1)
 
         # 构造输出频域张量
         out_ft = jnp.zeros((batch, self.out_channels, Nx, Ny//2+1), dtype=jnp.complex64)
